+++ cg/cli/upload/delivery_report.py
@@ -1,5 +1,4 @@
 """Code for uploading delivery report from the CLI"""
-import datetime as dt
 import logging
 import sys
 from pathlib import Path
@@ -10,9 +9,10 @@ from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.apps.scout.scoutapi import ScoutAPI
 from cg.constants import Pipeline
 from cg.exc import CgError, DeliveryReportError
+from cg.meta.report.api import ReportAPI
+from cg.meta.workflow.mip_dna import MipDNAAnalysisAPI
 
 from .utils import suggest_cases_delivery_report
-from cg.meta.workflow.mip_dna import MipDNAAnalysisAPI
 
 LOG = logging.getLogger(__name__)
 SUCCESS = 0
@@ -44,7 +44,7 @@ def delivery_reports(context, print_console, force_report):
 
             context.invoke(
                 delivery_report,
-                family_id=analysis_obj.family.internal_id,
+                case_id=analysis_obj.family.internal_id,
                 print_console=print_console,
                 force_report=force_report,
             )
@@ -79,7 +79,7 @@ def delivery_reports(context, print_console, force_report):
 
 
 @click.command("delivery-report")
-@click.argument("family_id", required=False)
+@click.argument("case_id", required=False)
 @click.option(
     "-p",
     "--print",
@@ -94,8 +94,17 @@ def delivery_reports(context, print_console, force_report):
     is_flag=True,
     help="overrule report validation",
 )
+@click.option(
+    "--analysis-started-at", help="Use the analysis started at (i.e.  '2020-05-28  12:00:46')"
+)
 @click.pass_context
-def delivery_report(context, family_id, print_console, force_report):
+def delivery_report(
+    context: click.Context,
+    case_id: str,
+    print_console: bool,
+    force_report: bool,
+    analysis_started_at: str = None,
+) -> None:
     """Generates a delivery report for a case and uploads it to housekeeper and scout
 
     The report contains data from several sources:
@@ -150,63 +159,44 @@ def delivery_report(context, family_id, print_console, force_report):
     click.echo(click.style("----------------- DELIVERY_REPORT -------------"))
 
     analysis_api: MipDNAAnalysisAPI = context.obj["analysis_api"]
+    report_api: ReportAPI = context.obj["report_api"]
 
-    def _add_delivery_report_to_hk(
-        delivery_report_file: Path, hk_api: HousekeeperAPI, family_id: str
-    ):
-        delivery_report_tag_name = "delivery-report"
-        version_obj = hk_api.last_version(family_id)
-        uploaded_delivery_report_files = hk_api.get_files(
-            bundle=family_id,
-            tags=[delivery_report_tag_name],
-            version=version_obj.id,
-        )
-        number_of_delivery_reports = len(uploaded_delivery_report_files.all())
-        is_bundle_missing_delivery_report = number_of_delivery_reports == 0
-
-        if is_bundle_missing_delivery_report:
-            file_obj = hk_api.add_file(
-                delivery_report_file.name, version_obj, delivery_report_tag_name
-            )
-            hk_api.include_file(file_obj, version_obj)
-            hk_api.add_commit(file_obj)
-            return file_obj
-
-        return None
-
-    def _update_delivery_report_date(status_api, case_id: str):
-        case_obj = status_api.family(case_id)
-        analysis_obj = case_obj.analyses[0]
-        analysis_obj.delivery_report_created_at = dt.datetime.now()
-        status_api.commit()
-
-    report_api = context.obj["report_api"]
-
-    if not family_id:
-        suggest_cases_delivery_report(context)
+    if not case_id or not analysis_api.status_db.family(internal_id=case_id):
+        suggest_cases_delivery_report(context, pipeline=Pipeline.MIP_DNA)
         context.abort()
 
+    if not analysis_started_at:
+        analysis_started_at = analysis_api.status_db.family(case_id).analyses[0].started_at
+
+    LOG.info("Using analysis started at: %s", analysis_started_at)
+
     if print_console:
-        delivery_report_html = report_api.create_delivery_report(family_id, force_report)
+        delivery_report_html = report_api.create_delivery_report(
+            case_id, analysis_started_at, force_report
+        )
         click.echo(delivery_report_html)
         return
 
     delivery_report_file = report_api.create_delivery_report_file(
-        family_id,
-        file_path=Path(analysis_api.root, family_id),
+        case_id,
+        file_path=Path(analysis_api.root, case_id),
         accept_missing_data=force_report,
+        analysis_date=analysis_started_at,
     )
-    added_file = _add_delivery_report_to_hk(
-        delivery_report_file, analysis_api.housekeeper_api, family_id
+
+    added_file = report_api.add_delivery_report_to_hk(
+        delivery_report_file, analysis_api.housekeeper_api, case_id, analysis_started_at
     )
 
     if added_file:
-        click.echo(click.style("uploaded to housekeeper", fg="green"))
+        click.echo(click.style("Uploaded to housekeeper", fg="green"))
     else:
-        click.echo(click.style("already uploaded to housekeeper, skipping", fg="yellow"))
+        click.echo(click.style("Already uploaded to housekeeper, skipping", fg="yellow"))
 
-    context.invoke(delivery_report_to_scout, case_id=family_id)
-    _update_delivery_report_date(analysis_api.status_db, family_id)
+    context.invoke(delivery_report_to_scout, case_id=case_id)
+    report_api.update_delivery_report_date(
+        analysis_api.status_db, case_id, analysis_date=analysis_started_at
+    )
 
 
 @click.command("delivery-report-to-scout")
@@ -216,7 +206,7 @@ def delivery_report(context, family_id, print_console, force_report):
     "--dry-run",
     "dry_run",
     is_flag=True,
-    help="run command without uploading to " "scout",
+    help="run command without uploading to scout",
 )
 @click.pass_context
 def delivery_report_to_scout(context, case_id: str, dry_run: bool):
@@ -224,7 +214,7 @@ def delivery_report_to_scout(context, case_id: str, dry_run: bool):
     analysis_api: MipDNAAnalysisAPI = context.obj["analysis_api"]
 
     if not case_id:
-        suggest_cases_delivery_report(context)
+        suggest_cases_delivery_report(context, pipeline=Pipeline.MIP_DNA)
         context.abort()
 
     uploaded_delivery_report_files = analysis_api.housekeeper_api.get_files(